Support for x64 (64 bit computers).
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@4674 640d5228-2204-0410-8d71-8f46fa6850b4

+++ ads2/hdr_api/getaddr.c
@@ -31,7 +31,7 @@ COPYRIGHT:	University Corporation for Atmospheric Research, 1992
 
 
 /* -------------------------------------------------------------------- */
-int GetPrimaryAddr(const char vn[], long *result)
+int GetPrimaryAddr(const char vn[], int32_t *result)
 {
   int	indx;
 
@@ -48,7 +48,7 @@ int GetPrimaryAddr(const char vn[], long *result)
 }	/* END GETADSADDRESS */
 
 /* -------------------------------------------------------------------- */
-int GetSecondaryAddr(const char vn[], long *result)
+int GetSecondaryAddr(const char vn[], int32_t *result)
 {
   int	indx;
